Add a way to get and set the logging level
In order for `Logging` to be swapped in for `logging.Logger`, certain
methods (beyond just the logging calls) need to be exposed.
`getEffectiveLevel` and `setLevel` are things that can be referenced
outside of Henson-Logging. `get_effective_level` and `set_level` are
being added as the canonical names, with `getEffectiveLevel` and
`setLevel` being added for API compatibility.